pyproject.toml
[tool.poetry]
name = "taskiq-nats"
version = "0.1.0"
description = ""
authors = ["Pavel Kirilin <win10@list.ru>"]
readme = "README.md"
packages = [{ include = "taskiq_nats" }]

[tool.poetry.dependencies]
python = "^3.9"
nats-py = "^2.2.0"
taskiq = "^0.2.0"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
pre-commit = "^3.1.1"
mypy = "^1.1.1"
flake8 = "^4.0.0"
wemake-python-styleguide = "^0.17.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
